+++ appseed/aura/aura/os/windows_common/windows_common_file.cpp
@@ -6,7 +6,7 @@
 
    hwstring hwPath(MAX_PATH * 16);
 
-   if (!GetModuleFileNameW(NULL, hwPath, hwPath.count()))
+   if (!GetModuleFileNameW(NULL, hwPath, (DWORD) hwPath.count()))
    {
 
       return "";
